+++ inst/app/tools/analysis/logistic_ui.R
@@ -632,7 +632,7 @@ logistic_report <- function() {
     pred_args <- clean_args(logit_pred_inputs(), logit_pred_args[-1])
 
     if (!radiant.data::is_empty(pred_args$pred_cmd)) {
-      pred_args$pred_cmd <- strsplit(pred_args$pred_cmd, ";")[[1]]
+      pred_args$pred_cmd <- strsplit(pred_args$pred_cmd, ";\\s*")[[1]]
     } else {
       pred_args$pred_cmd <- NULL
     }