+++ inst/app/tools/analysis/mnl_ui.R
@@ -633,7 +633,7 @@ mnl_report <- function() {
     pred_args <- clean_args(mnl_pred_inputs(), mnl_pred_args[-1])
 
     if (!radiant.data::is_empty(pred_args$pred_cmd)) {
-      pred_args$pred_cmd <- strsplit(pred_args$pred_cmd, ";")[[1]]
+      pred_args$pred_cmd <- strsplit(pred_args$pred_cmd, ";\\s*")[[1]]
     } else {
       pred_args$pred_cmd <- NULL
     }