+++ inst/app/tools/analysis/naivebayes_ui.R
@@ -461,7 +461,7 @@ nb_report <- function() {
     pred_args <- clean_args(nb_pred_inputs(), nb_pred_args[-1])
 
     if (!radiant.data::is_empty(pred_args$pred_cmd)) {
-      pred_args$pred_cmd <- strsplit(pred_args$pred_cmd, ";")[[1]]
+      pred_args$pred_cmd <- strsplit(pred_args$pred_cmd, ";\\s*")[[1]]
     } else {
       pred_args$pred_cmd <- NULL
     }