+++ inst/app/tools/analysis/nn_ui.R
@@ -596,7 +596,7 @@ nn_report <- function() {
     pred_args <- clean_args(nn_pred_inputs(), nn_pred_args[-1])
 
     if (!radiant.data::is_empty(pred_args$pred_cmd)) {
-      pred_args$pred_cmd <- strsplit(pred_args$pred_cmd, ";")[[1]]
+      pred_args$pred_cmd <- strsplit(pred_args$pred_cmd, ";\\s*")[[1]]
     } else {
       pred_args$pred_cmd <- NULL
     }