+++ inst/app/tools/analysis/rforest_ui.R
@@ -613,7 +613,7 @@ rf_report <- function() {
     pred_args <- clean_args(rf_pred_inputs(), rf_pred_args[-1])
 
     if (!radiant.data::is_empty(pred_args$pred_cmd)) {
-      pred_args$pred_cmd <- strsplit(pred_args$pred_cmd, ";")[[1]]
+      pred_args$pred_cmd <- strsplit(pred_args$pred_cmd, ";\\s*")[[1]]
     } else {
       pred_args$pred_cmd <- NULL
     }