+++ NEWS.md
@@ -1,3 +1,7 @@
+# radiant.model 1.4.9
+
+* Fix when parsing commands using strsplit on ';'
+
 # radiant.model 1.4.8
 
 * gsub("[\x80-\xFF]", "", text) is no longer valid in R 4.2.0 and above. Non-asci symbols will now be escaped using stringi